feat(HTD-70): move ui elements out to home and move to ui page
